% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.prime.R
\name{dp.bal}
\alias{dp.bal}
\alias{dp.estim}
\title{D' Balance}
\usage{
dp.bal(speed, cs = 2, d0 = 100, t.diff = 0.1, d.bal.min = -50)

dp.estim(speed, max.speed, crit.speed, duration)
}
\arguments{
\item{speed}{Athlete's speed}

\item{cs}{Critical speed}

\item{d0}{D'}

\item{t.diff}{Sample rate in seconds}
}
\description{
Returns an athlete's D' balance based on their current speed and critical speed. All arguments must be known to use this function.
}
\details{
If you do not know an athlete's critical speed or D', please use the `stamina.player.profile()` function.
}
\section{Functions}{
\itemize{
\item \code{dp.estim}: Estimate of an athlete's D'
}}

