% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.power.R
\name{met.power}
\alias{met.power}
\alias{met.power.update}
\alias{met.energy}
\alias{met.energy.update}
\alias{met.energy.kcal}
\title{Metabolic Energy and Power}
\usage{
met.power(speed, accel)

met.power.update(speed, accel)

met.energy(accel, kcal = FALSE)

met.energy.update(accel, kcal = FALSE)

met.energy.kcal(met.energy)
}
\arguments{
\item{speed}{Athlete's speed}

\item{accel}{Athlete's acceleration on the x-axis}
}
\description{
This function returns a player's instantaneous metabolic energy or power.
}
\section{Functions}{
\itemize{
\item \code{met.power.update}: Updated metabolic power algorithm

\item \code{met.energy}: Metabolic energy

\item \code{met.energy.update}: Updated metabolic energy

\item \code{met.energy.kcal}: Rough kcal estimate from an athlete's metabolic energy
}}

