% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.load.R
\name{pl.player.load}
\alias{pl.player.load}
\title{Player Load}
\usage{
pl.player.load(ax, ay, az, cumulative = FALSE, sample.rate = 100)
}
\arguments{
\item{ax}{Acceleration on the x-axis}

\item{ay}{Acceleration on the y-axis (optional)}

\item{az}{Acceleration on the z-axis (optional)}

\item{cumulative}{Summed player load over time}

\item{sample.rate}{Sample rate in Hz}
}
\description{
This function is built off Catapult Sports (R) player load model. Since some player tracking data returns a single acceleration vector, `ay` and `az` are optional.
}
\details{
`cumulative` returns a vector with the summation of the player load over time. Defualt is set to `FALSE`
}
