% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.as.R
\name{save.as.csv}
\alias{save.as.csv}
\alias{save.as.excel}
\alias{save.as.rdata}
\alias{save.as.rds}
\title{Save As}
\usage{
save.as.csv(data.set, file.name)

save.as.excel(data.set, file.name)

save.as.rdata(data.set, file.name)

save.as.rds(data.set, file.name)
}
\arguments{
\item{data.set}{Data frame containing stamina data}

\item{file.name}{Intended file name}
}
\description{
Save the output `data.frame`s as a .csv or .rds file
}
\section{Functions}{
\itemize{
\item \code{save.as.excel}: Alternate call for save.as.csv

\item \code{save.as.rdata}: Alternate call for save.as.rds

\item \code{save.as.rds}: Alternate call for save.as.rdata
}}

