% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamina.R
\name{stamina.player.profile}
\alias{stamina.player.profile}
\title{Stamina Player Profile}
\usage{
stamina.player.profile(speed, dur = 600, sample.rate = 10, algo = "slow")
}
\arguments{
\item{speed}{Athete's speed}

\item{dur}{Duration of the velocity-duration curve}

\item{sample.rate}{Sample rate in Hz}

\item{algo}{Speed of the function}
}
\description{
Returns approximations of an athlete's critical speed, D', max aerobic speed, and VO2 max.
}
\details{
`algo` is the algorithm used. Default is slow which can take 10 - 60 s to run, depending on the machine. When set to fast, the algorithm will return a rough estimate of an athlete's critical speed.

`dur` is set to 600 s (10 minutes). It is not advised to change this value.

**Note** The VO2 max algorithm is an approximation and has not been validated. Results are typically +/- 10%. If the data is very messy, results can be highly erratic.
}
