% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamina.R
\name{stamina.results.model}
\alias{stamina.results.model}
\title{Stamina Model Results}
\usage{
stamina.results.model(
  speed,
  ax,
  ay = NA,
  az = NA,
  vo2 = "auto",
  crit.speed = "auto",
  d.prime = "auto",
  player.load = TRUE,
  sample.rate = 10
)
}
\arguments{
\item{speed}{Athlete's speed}

\item{ax}{Athlete's acceleration on the x-axis}

\item{ay}{Athlete's acceleration on the y-axis (optional)}

\item{az}{Athlete's acceleration on the z-axis (optional)}

\item{vo2}{VO2 model}

\item{crit.speed}{Athlete's critical speed}

\item{d.prime}{Athlete's initial D' value}

\item{player.load}{If FALSE, overrides ax and sets it to zero}

\item{sample.rate}{Sample rate in Hz}
}
\description{
Returns a data.frame that includes their speed, acceleration, player load, cumulative player load, metabolic energy, metabolic power, V(dot)O2, approximate kcal expenditure, cumulative kcal expenditure, and D' balance.
}
\details{
If `crit.speed` or `d.prime` are unknown, set the argument to `auto` and the function will estimate these values. To return the estimated values, call `stamina.player.profile()`.

If `player.load` is set to `FALSE`, the function assumes that acceleration is not provided and will override values in the `ax` argument. You can omit the `ax` argument to achieve the same results.

`vo2` can take on one of the following: `"vo2.walk"`, `"vo2.jog"`, `"vo2.track"`, `"vo2.treadmill"`. The default is set to `"vo2.track"` with the assumption that is the nearest approximation to actual V(dot)O2.
}
