% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vo2.fns.R
\name{vo2.auto}
\alias{vo2.auto}
\alias{vo2.walk}
\alias{vo2.jog}
\alias{vo2.track}
\alias{vo2.treadmill}
\alias{vo2.kcal}
\alias{vo2.kcal.dist}
\title{VO2 Max}
\usage{
vo2.auto(speed.kph, vo2.max = 60)

vo2.walk(speed.kph)

vo2.jog(speed.kph)

vo2.track(speed.kph)

vo2.treadmill(speed.kph)

vo2.kcal(speed.kph)

vo2.kcal.dist(speed.kph, sample.rate = 10)
}
\arguments{
\item{speed.kph}{The player's speed in kph}
}
\description{
Estimates an athlete's V(dot)O2. `vo2.auto` selects the most appropriate algorithm based on the athlete's speed. VO2 max is capped at 60 ml/kg/minute if it isn't specified.
}
\details{
Algorithms are built for individuals who are walking (< 10 kph), jogging (10 to 17 kph), running on a track (8 to 25 kph), and running on a treadmill (8 to 25 kph).

To convert m/s to kps, multiple by 3.6

VO2 is returned in ml/kg/min
}
\section{Functions}{
\itemize{
\item \code{vo2.walk}: VdotO2 at a walking pace

\item \code{vo2.jog}: VO2 at a jogging pace

\item \code{vo2.track}: VO2 while running on a track

\item \code{vo2.treadmill}: VO2 while running on a treadmill

\item \code{vo2.kcal}: The kcal equivalent expenditure in kcal/kg/km

\item \code{vo2.kcal.dist}: The kcal equivalent expenditure for a distance travelled
}}

