% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vo2.fns.R
\name{vo2.estim}
\alias{vo2.estim}
\alias{vo2.estim.2}
\alias{vo2.estim.3}
\title{VO2 Max Estimates}
\usage{
vo2.estim(MAS)

vo2.estim.2(MAS)

vo2.estim.3(MAS)
}
\arguments{
\item{MAS}{Max aerobic speed}
}
\description{
An estimation of an athlete's VO2 max in ml/kg/min
}
\section{Functions}{
\itemize{
\item \code{vo2.estim.2}: Estimate of an athlete's VO2 max v.2

\item \code{vo2.estim.3}: Estimate of an athlete's VO2 max v.3
}}

