% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vo2.R
\name{vo2.model.fit}
\alias{vo2.model.fit}
\alias{vo2.model.integral}
\title{VO2 Max Model Fit}
\usage{
vo2.model.fit(alpha, beta, theta, duration, cf = 1.25)

vo2.model.integral(alpha, beta, theta, duration, cf = 1.25)
}
\arguments{
\item{alpha}{alpha}

\item{beta}{beta}

\item{theta}{theta}

\item{duration}{duration}

\item{cf}{correction factor}
}
\description{
VO2 Max Model Fit
}
\section{Functions}{
\itemize{
\item \code{vo2.model.integral}: Returns the integral of the velocity duration curve
}}

